elf: Remove -fno-tree-loop-distribute-patterns usage on dl-support
Besides the option being gcc specific, this approach is still fragile
and not future proof since we do not if this will be the only
optimization option gcc will add that transform loops to memset
(or any libcall).

This patch add a new header dl-symbol-hacks.h that can be used to
redirect the compiler generated libcalls to port specific string
implementatation if required (defined by dl-ifunc-generic.h).

Checked on x86_64-linux-gnu and aarch64-linux-gnu.